Added regression test for bug with ABI start
Added a regression test for a bug with the fake ABI start times.  ABI's
exists method is too simple, it needs to search for a bunch of ABI files
and then decide if any of them *covers* the requested time.  The test is
currently failing as this is not currently implemented.