Fix functionality in ABI.exists
The ABI.exists functionality was totally broken, looking only for an
exact start time match.  Now it uses satpy.readers.find_files_and_readers
to search for files covering a certain timestamp, through either the
present directory or a directory covering 15 minutes earlier, and if it
is requested to need the past it assures also that t-60 minutes exists
as well as either t-30 or t-20 minutes.  This is now mostly covered by
unittests except for some of the error checks (5 lines missing).