Improve automated ABI downloading
Add a function that downloads ABI data for an arbitrary period, rather
than for a day from midnight to midnight.  Use this to more
intelligently download ABI data from t-65 to t+5 minutes when
unavailable.  Adapt tests accordingly.  This depends on a recent (still
to be reviewed) improvement in Pytroll to search for files on a remote
filesystem.

Also improve the logging in the db, by having more logging statements
and giving each db element a name.