Bugfix in ABI downloading
In the fog DB, ABI was downloaded to a different location than where it
was being looked for by .exists().  This means that firstly it kept
retrying to download it (although when downloading finding it was
already there), secondly that the files were subsequently not found when
passed to a Scene.