Handle duplicate lat/lons better
Handle duplicate lat/lons better, we don't want to extract the same
lat/lon from the same file twice.  Also, when multpile datasets in the
same scene have different start times, issue a warning and get on with
it, don't crash and certainly don't enter them with different times in
the database.